Test split in two to fix ibm-ws-bnd.xml update issue
