+++ dev/io.openliberty.jaxws.security_fat/fat/src/com/ibm/ws/jaxws/security/fat/EJBInJarServiceSecurityTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2023 IBM Corporation and others.
+ * Copyright (c) 2023, 2024 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -21,44 +21,31 @@ import java.net.ProtocolException;
 import java.net.URL;
 
 import org.jboss.shrinkwrap.api.spec.WebArchive;
-import org.junit.After;
-import org.junit.AfterClass;
 import org.junit.Assert;
-import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Rule;
-import org.junit.Test;
 import org.junit.rules.TestName;
-import org.junit.runner.RunWith;
 
 import com.ibm.websphere.simplicity.ShrinkHelper;
 import com.ibm.websphere.simplicity.log.Log;
 
-import componenttest.annotation.AllowedFFDC;
 import componenttest.annotation.Server;
-import componenttest.custom.junit.runner.FATRunner;
-import componenttest.custom.junit.runner.Mode;
 import componenttest.topology.impl.LibertyServer;
 import componenttest.topology.utils.HttpUtils;
 
 /**
- * This class tests the security of ejb based web services, which are packaged
- * in a jar package. - no security constraints - security constraints are
- * configured in ibm-ws-bnd.xml
+ * This abstract class provides a super class for two tests split by their usage of ibm-ws.bnd.xml or not
  */
-@RunWith(FATRunner.class)
-public class EJBInJarServiceSecurityTest {
+public abstract class EJBInJarServiceSecurityTest {
 
     @Server("EJBInJarSecurityServer")
     public static LibertyServer server;
 
-    private final static int REQUEST_TIMEOUT = 10;
-
     @Rule
     public TestName testName = new TestName();
 
-    @BeforeClass
-    public static void setUp() throws Exception {
+    private final static int REQUEST_TIMEOUT = 10;
+
+    protected static void init() throws Exception {
         WebArchive war = ShrinkHelper.buildDefaultApp("EJBInJarServiceSecurityClient", "com.ibm.samples.jaxws",
                                                       "com.ibm.samples.servlet");
 
@@ -67,112 +54,20 @@ public class EJBInJarServiceSecurityTest {
 
         ExplodedShrinkHelper.explodedJarToDestination(server, "apps/EJBInJarServiceSecurity.ear",
                                                       "EJBInJarServiceSecurity", "com.ibm.samples.jaxws", "com.ibm.sample.ejb");
-
-        // Make sure we don't fail because we try to start an
-        // already started server
-        try {
-            server.startServer("WebServiceRefTest.log");
-        } catch (Exception e) {
-            System.out.println(e.toString());
-        }
-
-        // Pause for application to start successfully
-        server.waitForStringInLog("CWWKZ0001I.*EJBInJarServiceSecurity");
-
-        // BASE_URL = "http://" + server.getHostname() + ":" +
-        // server.getHttpDefaultPort();
     }
 
-    @AfterClass
-    public static void tearDown() throws Exception {
-        if (server != null && server.isStarted()) {
-            server.stopServer();
-        }
-    }
-
-    @Before
-    public void beforeEachTest() throws Exception {
-
-    }
-
-    @After
-    public void afterEachTest() throws Exception {
-        if (server != null && server.isStarted()) {
-            server.stopServer();
-        }
-    }
-
-    @Mode(componenttest.custom.junit.runner.Mode.TestMode.FULL)
-    @Test
-    @AllowedFFDC({ "java.rmi.AccessException" })
-    public void test_ejbws_in_jar_security_with_bndfile() throws Exception {
-
-        // add ibm-ws-bnd.xml
-        if (!server.fileExistsInLibertyServerRoot(
-                                                  "apps/EJBInJarServiceSecurity.ear/EJBInJarServiceSecurity.jar/META-INF/ibm-ws-bnd.xml")) {
-            server.copyFileToLibertyServerRoot(
-                                               "apps/EJBInJarServiceSecurity.ear/EJBInJarServiceSecurity.jar/META-INF/ibm-ws-bnd.xml",
-                                               "EJBWSSecurityFileStore/ibm-ws-bnd.xml");
-        }
-
+    public static void startServer(String logName) throws Exception {
         if (server != null && !server.isStarted()) {
-            // delete whatever ejb jar
-            server.startServer();
-
+            server.startServer(logName);
             checkAppsReady();
             server.setMarkToEndOfLog();
         }
+    }
 
-        runTest("user1", "user2pwd", "SayHelloService", "Hello user1 from ejb web service.", false);
-        runTest("user1", "user1pwd", "SayHelloService", "Hello user1 from ejb web service.", false);
-        runTest("user2", "user2pwd", "SayHelloService", "Hello user2 from ejb web service.", true);
-        runTest("user3", "user3pwd", "SayHelloService", "Hello user3 from ejb web service.", true);
-        runTest("user4", "user4pwd", "SayHelloService", "Hello user4 from ejb web service.", false);
-
-        runTest("user1", "user2pwd", "SecuredSayHelloService", "Hello user1 from secured ejb web service.", false);
-        runTest("user1", "user1pwd", "SecuredSayHelloService", "Hello user1 from secured ejb web service.", false);
-        runTest("user2", "user2pwd", "SecuredSayHelloService", "Hello user2 from secured ejb web service.", true);
-        runTest("user3", "user3pwd", "SecuredSayHelloService", "Hello user3 from secured ejb web service.", false);
-        runTest("user4", "user4pwd", "SecuredSayHelloService", "Hello user4 from secured ejb web service.", false);
-
+    public static void stopServer() throws Exception {
         if (server != null && server.isStarted()) {
             server.stopServer();
         }
-        // Adding another deletion here to be sure deletion happens before test_ejbws_in_jar_security_without_bndfile test
-        server.deleteFileFromLibertyServerRoot(
-                                               "apps/EJBInJarServiceSecurity.ear/EJBInJarServiceSecurity.jar/META-INF/ibm-ws-bnd.xml");
-    }
-
-    @Mode(componenttest.custom.junit.runner.Mode.TestMode.FULL)
-    @Test
-    @AllowedFFDC({ "java.rmi.AccessException" })
-    public void test_ejbws_in_jar_security_without_bndfile() throws Exception {
-
-        // delete ibm-ws-bnd.xml
-        if (server.fileExistsInLibertyServerRoot(
-                                                 "apps/EJBInJarServiceSecurity.ear/EJBInJarServiceSecurity.jar/META-INF/ibm-ws-bnd.xml"))
-            server.deleteFileFromLibertyServerRoot(
-                                                   "apps/EJBInJarServiceSecurity.ear/EJBInJarServiceSecurity.jar/META-INF/ibm-ws-bnd.xml");
-
-        if (server != null && !server.isStarted()) {
-            // delete whatever ejb jar
-            server.startServer();
-
-            checkAppsReady();
-            server.setMarkToEndOfLog();
-        }
-
-        runTest("user1", "user2pwd", "SayHelloService", "Hello user1 from ejb web service.", true);
-        runTest("user1", "user1pwd", "SayHelloService", "Hello user1 from ejb web service.", true);
-        runTest("user2", "user2pwd", "SayHelloService", "Hello user2 from ejb web service.", true);
-        runTest("user3", "user3pwd", "SayHelloService", "Hello user3 from ejb web service.", true);
-        runTest("user4", "user4pwd", "SayHelloService", "Hello user4 from ejb web service.", true);
-
-        runTest("user1", "user2pwd", "SecuredSayHelloService", "Hello user1 from secured ejb web service.", false);
-        runTest("user1", "user1pwd", "SecuredSayHelloService", "Hello user1 from secured ejb web service.", true);
-        runTest("user2", "user2pwd", "SecuredSayHelloService", "Hello user2 from secured ejb web service.", true);
-        runTest("user3", "user3pwd", "SecuredSayHelloService", "Hello user3 from secured ejb web service.", false);
-        runTest("user4", "user4pwd", "SecuredSayHelloService", "Hello user4 from secured ejb web service.", false);
     }
 
     protected void runTest(String username, String password, String serviceName, String responseString,
@@ -196,7 +91,7 @@ public class EJBInJarServiceSecurityTest {
         }
     }
 
-    private void checkAppsReady() {
+    private static void checkAppsReady() {
         Assert.assertNotNull("The application EJBInJarServiceSecurity did not appear to have started",
                              server.waitForStringInLog("CWWKZ0001I.*EJBInJarServiceSecurity"));
         Assert.assertNotNull("Security service did not report it was ready", server.waitForStringInLog("CWWKS0008I"));
