+++ dev/io.openliberty.jaxws.security_fat/fat/src/com/ibm/ws/jaxws/security/fat/FATSuite.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2023 IBM Corporation and others.
+ * Copyright (c) 2023, 2024 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
@@ -24,7 +24,8 @@ import componenttest.rules.repeater.RepeatTests;
  */
 @RunWith(Suite.class)
 @SuiteClasses({ BasicAuthWithoutSSLTest.class,
-                EJBInJarServiceSecurityTest.class,
+                EJBInJarServiceSecurityWithBndTest.class,
+                EJBInJarServiceSecurityWithoutBndTest.class,
                 EJBInWarServiceSecurityTest.class,
                 POJOServiceSecurityTest.class,
                 WsdlLocationHttpsTest.class
