--- dev/io.openliberty.jaxws.security_fat/test-applications/EJBInJarServiceSecurity/resources/META-INF/ibm-ws-bnd.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<webservices-bnd xmlns="http://websphere.ibm.com/xml/ns/javaee" 
-                      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" 
-                      xsi:schemaLocation="http://websphere.ibm.com/xml/ns/javaee http://websphere.ibm.com/xml/ns/javaee/ibm-ws-bnd_1_0.xsd" version="1.0">
-                      
-    <http-publishing>
-        <webservice-security> 
-            <security-constraint>
-                    <web-resource-collection>
-                        <web-resource-name>Resource for SayHelloService</web-resource-name>
-                        <url-pattern>/SayHelloService</url-pattern>
-                    </web-resource-collection>
-                    <auth-constraint>
-                        <description>Only role23 is allowed</description>
-                        <role-name>role23</role-name>
-                    </auth-constraint>
-                </security-constraint>        
-                <security-constraint>
-                    <web-resource-collection>
-                        <web-resource-name>Resource for SecuredSayHelloService</web-resource-name>
-                        <url-pattern>/SecuredSayHelloService</url-pattern>
-                    </web-resource-collection>
-                    <auth-constraint>
-                        <description>Only role23 is allowed</description>
-                        <role-name>role23</role-name>
-                    </auth-constraint>
-                </security-constraint>                      
-               <!-- SECURITY ROLES -->
-               <security-role id="role23_id">
-                    <role-name>role23</role-name>
-               </security-role>
-               <!-- AUTHENTICATION METHOD: Basic authentication -->
-               <login-config>
-                    <auth-method>BASIC</auth-method>
-               </login-config>
-        </webservice-security>
-    </http-publishing>
-</webservices-bnd>
